requirements.txt
ebook-searching-proto==0.1.3
grpcio==1.68.0
protobuf==3.20.3
